Update the framework

+++ MinimalApi/Endpoints/PostEndPointDefinitions.cs
@@ -24,27 +24,35 @@ public class PostEndPointDefinitions : IEndpointDefiniton
 
     private async Task<IResult> GetPostById(IPostService _postService, int id)
     {
-        var post = await _postService.GetPost(id);
-        return TypedResults.Ok(post);
+        return await _postService.GetPost(id) is Post post
+            ? TypedResults.Ok(post)
+            : TypedResults.NotFound();
     }
 
     private async Task<IResult> CreatePost(IPostService _postService, Post post)
     {
+        if (string.IsNullOrEmpty(post.Content)) return TypedResults.BadRequest();
+
         var createdPost = await _postService.CreatePost(post);
         return Results.CreatedAtRoute("GetPostById", new { createdPost.Id }, createdPost);
     }
 
     private async Task<IResult> UpdatePost(IPostService _postService, int id, Post post)
     {
-        if (string.IsNullOrEmpty(post.Content))
-        { return TypedResults.BadRequest(); }
+        if (string.IsNullOrEmpty(post.Content)) return TypedResults.BadRequest();
 
-        var updatedPost = await _postService.UpdatePost(id, post.Content);
-        return TypedResults.Ok(updatedPost);
+        var existingPost = await _postService.GetPost(id);
+        if (existingPost is null) return TypedResults.NotFound();
+
+        await _postService.UpdatePost(id, post.Content);
+
+        return TypedResults.NoContent();
     }
 
     private async Task<IResult> DeletePost(IPostService _postService, int id)
     {
+        if (await _postService.GetPost(id) is not Post post) return TypedResults.NotFound();
+
         await _postService.DeletePost(id);
         return TypedResults.NoContent();
     }
